fix: restore dist main build output
