+++ package.json
@@ -19,7 +19,7 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --passWithNoTests",
-    "migrate:sqlite-to-pg": "ts-node --transpile-only scripts/migrate-sqlite-to-pg.ts"
+    "migrate:sqlite-to-pg": "ts-node --project tsconfig.scripts.json --transpile-only scripts/migrate-sqlite-to-pg.ts"
   },
   "dependencies": {
     "@nestjs/axios": "^3.0.0",
